+++ src/models/evaluation.ts
@@ -1,4 +1,11 @@
-export interface evaluatedUser {
-  evaluatee: string;
+export interface apiEvaluatedUser {
+  projectId: number;
+  evaluateeId: number;
   score: number[];
+}
+
+export interface apiMemberList {
+  userId: number;
+  nickname: string;
+  evaluated: boolean;
 }
